gcs_to_bq_loader/requirements.txt
google-cloud-storage
google-cloud-bigquery 